--- Project02Eclipse/src/com/hemebiotech/analytics/Test/exceptions/FileNotFound.java
@@ -1,6 +1,0 @@
-package com.hemebiotech.analytics.Test.exceptions;
-
-public class FileNotFound extends Exception {
-
-
-}
