+++ Project02Eclipse/src/com/hemebiotech/analytics/packages/exceptions/ProgramException.java
@@ -1,0 +1,13 @@
+package com.hemebiotech.analytics.packages.exceptions;
+
+public class ProgramException extends Exception {
+
+    public void wrongPath () {
+
+    }
+
+    public void writeError () {
+
+    }
+
+}
